feat: support children
