+++ src/dropdown/Dropdown.tsx
@@ -3,12 +3,13 @@ import classNames from 'classnames';
 import omit from 'lodash/omit';
 import { DropdownOption, TdDropdownProps } from './type';
 import { ClassName } from '../common';
-import useConfig from '../hooks/useConfig';
 import Popup, { PopupVisibleChangeContext } from '../popup';
-import DropdownSub from './DropdownSub';
 import DropdownMenu from './DropdownMenu';
 import DropdownItem from './DropdownItem';
 import { dropdownDefaultProps } from './defaultProps';
+
+import useConfig from '../hooks/useConfig';
+import useDropdownOptions from './hooks/useDropdownOptions';
 
 export interface DropdownProps extends TdDropdownProps {
   className?: ClassName;
@@ -16,34 +17,44 @@ export interface DropdownProps extends TdDropdownProps {
 }
 
 const Dropdown = (props: DropdownProps) => {
-  const { popupProps = {}, disabled, placement, trigger, className, children, hideAfterItemClick } = props;
+  const {
+    popupProps = {},
+    disabled,
+    placement,
+    trigger,
+    className,
+    children,
+    hideAfterItemClick,
+    options: propsOptions,
+  } = props;
+
   const content = null;
   const arrayChildren = React.Children.toArray(children);
-
   const { classPrefix } = useConfig();
   const [isPopupVisible, togglePopupVisible] = useState(false);
   const dropdownClass = `${classPrefix}-dropdown`;
 
+  const options = useDropdownOptions(arrayChildren, propsOptions);
+
   const handleMenuClick = (data: DropdownOption, context: { e: React.MouseEvent<HTMLDivElement, MouseEvent> }) => {
-    if (hideAfterItemClick) {
-      togglePopupVisible(false);
-    }
+    if (hideAfterItemClick) togglePopupVisible(false);
+
     props?.onClick?.(data, context);
   };
-
-  const DropdownContent = () => <DropdownSub {...props} onClick={handleMenuClick}></DropdownSub>;
 
   const handleVisibleChange = (visible: boolean, context: PopupVisibleChangeContext) => {
     togglePopupVisible(visible);
     popupProps?.onVisibleChange?.(visible, context);
   };
 
+  const renderContent = <DropdownMenu {...props} options={options} onClick={handleMenuClick} />;
+
   const popupParams = {
     disabled,
     placement,
     trigger,
     showArrow: false,
-    content: content || DropdownContent(),
+    content: content || renderContent,
     ...omit(popupProps, 'onVisibleChange'),
     overlayInnerClassName: classNames(dropdownClass, className, popupProps?.overlayInnerClassName),
   };
