+++ src/dropdown/DropdownItem.tsx
@@ -22,29 +22,29 @@ const DropdownItem = forwardRef((props: DropdownItemProps, ref: React.RefObject<
   const { classPrefix } = useConfig();
   const maxColWidth = Number(maxColumnWidth) - 28;
   const minColWidth = Number(minColumnWidth) - 28;
-  const keyOnly = String(value);
   const prefixIconNode = <div className={`${classPrefix}-dropdown__item-icon`}>{prefixIcon}</div>;
   const dropdownItemRef = useRef(null);
 
   useRipple(ref || dropdownItemRef);
 
   return (
-    <li
-      key={keyOnly}
-      className={classNames(className, `${classPrefix}-dropdown__item--theme-${theme}`, {
-        [`${classPrefix}-dropdown__item--active`]: active,
-        [`${classPrefix}-dropdown__item--disabled`]: disabled,
-      })}
-      onClick={onClick ? (e) => onClick(value, e) : () => null}
-      style={{
-        maxWidth: `${maxColWidth}px`,
-        minWidth: `${minColWidth}px`,
-      }}
-      ref={ref || dropdownItemRef}
-    >
-      {prefixIcon && prefixIconNode}
-      {children}
-    </li>
+    <>
+      <li
+        className={classNames(className, `${classPrefix}-dropdown__item--theme-${theme}`, {
+          [`${classPrefix}-dropdown__item--active`]: active,
+          [`${classPrefix}-dropdown__item--disabled`]: disabled,
+        })}
+        onClick={onClick ? (e) => onClick(value, e) : () => null}
+        style={{
+          maxWidth: `${maxColWidth}px`,
+          minWidth: `${minColWidth}px`,
+        }}
+        ref={ref || dropdownItemRef}
+      >
+        {prefixIcon && prefixIconNode}
+        {children}
+      </li>
+    </>
   );
 });
